requirements.txt
Click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.5
Werkzeug==2.2.2
python-dotenv==0.21.0